requirements.txt
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
dj-database-url==0.5.0
Django==3.1.1
django-heroku==0.3.1
djangorestframework==3.12.1

gunicorn==20.0.4
isort==5.5.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.3

psycopg2==2.8.6
public==2019.4.13
pylint==2.6.0
pytz==2020.1
query-string==2019.4.13

six==1.15.0
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.2.0
wrapt==1.12.1
